Nginx fails with proposed key cipher and size.
This fails on the latest nginx with message stating the key is too small.

This is inline with the following wso2 docs

https://apim.docs.wso2.com/en/latest/install-and-setup/setup/reference/supported-cipher-suites/

Which state that this algorithm is deprecated and insecure. Increasing to 2048 seems to solve the issue, but this should be elevated to security for reviewal and consideration of other ciphers.